requirements.txt
agentql
playwright
python-dotenv
